Refactored BookmarkForm
