+++ src/shared/components/helpers/index.js
@@ -2,5 +2,6 @@ export { default as LinkTab } from "./LinkTab";
 export { default as SearchInput } from "./SearchInput";
 export { default as SnackbarCloseButton } from "./SnackbarCloseButton";
 export { default as StyledTreeItem } from "./StyledTreeItem";
+export { default as SubmitButton } from "./SubmitButton";
 export { default as TabPanel } from "./TabPanel";
 export { default as TreeItemStyles } from "./TreeItemStyles";
